FBP-61. Separate package for task-alike elements
