+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/tasks/BpmnDecisionTask.kt
@@ -1,7 +1,8 @@
-package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.tasks
 
 import com.github.pozo.KotlinBuilder
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.BpmnElementId
+import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.WithBpmnId
 
 /**
  * Is created from BpmnServiceTask based on its type
