+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/tasks/BpmnMuleTask.kt
@@ -1,19 +1,20 @@
-package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.tasks
 
 import com.github.pozo.KotlinBuilder
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.BpmnElementId
+import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.WithBpmnId
 
+/**
+ * Is created from BpmnServiceTask based on its type
+ */
 @KotlinBuilder
-data class BpmnBusinessRuleTask(
+data class BpmnMuleTask(
         override val id: BpmnElementId,
         val name: String?,
         val documentation: String?,
         val async: Boolean?,
-        val isForCompensation: Boolean?,
-        val ruleVariablesInput: String?,
-        val rules: String?,
-        val resultVariable: String?,
-        val exclude: Boolean?
+        val exclusive: Boolean?,
+        val isForCompensation: Boolean?
 ): WithBpmnId {
 
     override fun updateBpmnElemId(newId: BpmnElementId): WithBpmnId {
