+++ xml-parser-api/src/main/kotlin/com/valb3r/bpmn/intellij/plugin/bpmn/api/bpmn/elements/tasks/BpmnScriptTask.kt
@@ -1,22 +1,19 @@
-package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements
+package com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.tasks
 
 import com.github.pozo.KotlinBuilder
 import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.BpmnElementId
+import com.valb3r.bpmn.intellij.plugin.bpmn.api.bpmn.elements.WithBpmnId
 
 @KotlinBuilder
-data class BpmnUserTask(
+data class BpmnScriptTask(
         override val id: BpmnElementId,
         val name: String?,
         val documentation: String?,
         val async: Boolean?,
         val isForCompensation: Boolean?,
-        val assignee: String?,
-        val dueDate: String?,
-        val category: String?,
-        val formKey: String?,
-        val formFieldValidation: Boolean?,
-        val priority: String?,
-        val skipExpression: String?
+        val scriptBody: String?,
+        val scriptFormat: String?,
+        val autoStoreVariables: Boolean?
 ): WithBpmnId {
 
     override fun updateBpmnElemId(newId: BpmnElementId): WithBpmnId {
